Move metadata extraction into a library function
